\name{circle_mQTL}
\alias{circle_mQTL}

\title{
Circular genome-metabolome plot for mQTL.NMR
}
\description{
shows mQTL locations and relations with the metabolome on a central chemical axis
}
\usage{
circle_mQTL(results, Th = 0, chr = 9, spacing = 25)
}

\arguments{
  \item{results}{
a list containing mQTL mapping results generated by mQTL.NMR package
}
  \item{Th}{
a numerical parameter specifying LOD threshold
}
  \item{chr}{
a numerical value defining the chromosomes to show if necessary 
}
  \item{spacing}{
a numerical parameter specifiying the sapcing between chromosomes on the circular genome
}
}

\value{
A circuler plot where the central horizontal line corresponds to the NMR chemical axis, the circle represents the chromosomal positions, and the colored lines significant association between a shift and genomic location. 
}

\author{
Lyamine Hedjazi
}

\seealso{
\code{\link{pplot}}
}
\examples{
load_datafiles()
load(results)

circle_mQTL(results, Th=8,spacing=0)
}

\keyword{ function }

