\name{load_datafiles}
\alias{load_datafiles}

\title{
Load data files for examples
}
\description{
Data files are downloaded from the extdata directory to the user's working directory.  
}
\usage{
load_datafiles()
}

\value{
Loaded data files concern four datasets: raw metabolomic data ('phenofile.txt'), genomic data ('genofile.txt'), additional data ('physiodat.txt').     
}

\author{
Lyamine Hedjazi
}

\seealso{
\code{\link{format_mQTL}}
}
\examples{
# Load data files
 load_datafiles()
}

\keyword{ function }

